add forge building
this is a special type of build jobs where the command line is fixed.
it builds a base container with no input in the builder queue, called
forges, and allows other runner rules to use these forge containers
to run other jobs, such as building containers.

it is only for bootstraping the build process with a standard build
job sandbox. forges will not only contain build tools, but not
library/application/build target dependencies, and hence does not
need input.