config: add a cache readonly mark for PR configs (#71)
we will stop using "if it is a PR" for checking.

Rather, simply use if it runs on the premerge pipeline or the postmerge
pipeline. this will enable people to manually run and also for insiders
to run on non-master but still has the power to write to caching infra
and/or perform releases.